cast to int

+++ corehq/apps/reports/util.py
@@ -426,7 +426,7 @@ def batch_qs(qs, num_batches=10):
                 print article.body
     """
     total = qs.count()
-    batch_size = total / num_batches or 1
+    batch_size = int(total / num_batches) or 1
     for start in range(0, total, batch_size):
         end = min(start + batch_size, total)
         yield (start, end, total, qs[start:end])
